+++ apibuilder4elastic/custom_flow_nodes/api-builder-plugin-axway-api-management/src/actions.js
@@ -182,7 +182,8 @@ async function lookupAPIDetails(params, options) {
 	if(proxies == undefined) {
 		// To lookup the API in API-Manager the API-Name is required
 		if (!apiName) {
-			throw new Error(`API not configured locally, based on path: ${apiPath}. The API cannot be queried at the API Manager as no API name is given. Please configure this API path locally.`);
+			logger.info(`API not configured locally, based on path: ${apiPath}. The API cannot be queried at the API Manager as no API name is given. Please configure this API path locally.`);
+			return { name: 'Unknown API', method: 'Unknown Method'};
 		}
 		logger.debug(`API not configured locally, trying to get details from API-Manager.`);
 		proxies = await _getAPIProxy(apiName, groupId, params.region);
@@ -190,7 +191,8 @@ async function lookupAPIDetails(params, options) {
 		logger.info(`API-Details for API with path: '${apiPath}' looked up locally.`);
 	}
 	if(!proxies || proxies.length == 0) {
-		throw new Error(`No APIs found with name: '${apiName}'`);
+		logger.warn(`No APIs found with name: '${apiName}'`);
+		return { name: 'Unknown API', method: 'Unknown Method'};
 	}
 	var apiProxy = undefined;
 	for (i = 0; i < proxies.length; i++) {
@@ -201,7 +203,8 @@ async function lookupAPIDetails(params, options) {
 		}
 	}
 	if(!apiProxy) {
-		throw new Error(`No APIs found with name: '${apiName}' and apiPath: '${apiPath}'`);
+		logger.warn(`No APIs found with name: '${apiName}' and apiPath: '${apiPath}'`);
+		return { name: 'Unknown API', method: 'Unknown Method'};
 	}
 	// Skip  the lookups if the API is locally configured and just take it as it is
 	if(!apiProxy.locallyConfigured) {