+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Updated Elastic-Stack from version 7.12.1 to 7.14.0
 - Updated API-Builder version changed from Quezon to Timbuktu
+- API-Builder now always returns an Unknown API object if an API could not be found [#128](https://github.com/Axway-API-Management-Plus/apigateway-openlogging-elk/issues/128)
 
 ## [3.3.1] 2021-08-17
 ### Fixed